// client/src/api/generatedClient.ts
export interface DispatchDto {
    id: string;
    productId: string;
    productReference: string;
    productName: string;
    status: string;
    requestedBy: string;
    workshop: string;
    requestedAt: string;
    completedAt: string | undefined;
}